requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.5
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
Pillow==10.1.0
pytz==2025.2

